@@ -10,17 +10,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.9",
+    "@angular/animations": "^18.0.6",
     "@angular/cdk": "^17.3.9",
-    "@angular/common": "^17.3.9",
-    "@angular/compiler": "^17.3.9",
-    "@angular/core": "^17.3.9",
-    "@angular/forms": "^17.3.9",
-    "@angular/localize": "^17.3.9",
+    "@angular/common": "^18.0.6",
+    "@angular/compiler": "^18.0.6",
+    "@angular/core": "^18.0.6",
+    "@angular/forms": "^18.0.6",
+    "@angular/localize": "^18.0.6",
     "@angular/material": "^17.3.9",
-    "@angular/platform-browser": "^17.3.9",
-    "@angular/platform-browser-dynamic": "^17.3.9",
-    "@angular/router": "^17.3.9",
+    "@angular/platform-browser": "^18.0.6",
+    "@angular/platform-browser-dynamic": "^18.0.6",
+    "@angular/router": "^18.0.6",
     "@ng-bootstrap/ng-bootstrap": "^16.0.0",
     "@ngneat/until-destroy": "^10.0.0",
     "@popperjs/core": "^2.11.8",
@@ -50,9 +50,9 @@
     "zone.js": "^0.14.6"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.7",
-    "@angular/cli": "~17.3.7",
-    "@angular/compiler-cli": "^17.3.9",
+    "@angular-devkit/build-angular": "^18.0.7",
+    "@angular/cli": "~18.0.7",
+    "@angular/compiler-cli": "^18.0.6",
     "@types/jasmine": "~5.1.4",
     "@types/stripe": "^8.0.417",
     "@types/stripe-checkout": "^1.0.7",
